+++ src/pages/gongguPages/GongguMapPage.tsx
@@ -3,7 +3,7 @@ import { useState, useEffect } from "react";
 import { useLocation } from "react-router-dom";
 
 import { fetchWithAuth } from "../../utils/FetchWithAuth";
-import useDeviceSize from "../../useDeviceSize";
+import useDeviceSize from "../../hooks/useDeviceSize";
 import MapHeader from "../../components/gongguPages/gongguMapPage/MapHeader";
 import MapMenu from "../../components/gongguPages/gongguMapPage/MapMenu";
 import GongguMap from "../../components/gongguPages/gongguMapPage/GongguMap";
