+++ src/pages/shoppingPages/ShopItemPage.tsx
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import Header from "../../components/shoppingPages/ShopItemPage/Header.tsx";
 import Nav from "../../components/shoppingPages/ShopItemPage/Nav.tsx";
 import Main from "../../components/shoppingPages/ShopItemPage/Main.tsx";
-import useDeviceSize from "../../useDeviceSize.tsx";
+import useDeviceSize from "../../hooks/useDeviceSize.tsx";
 import { useState, useEffect } from "react";
 import { fetchWithAuth } from "../../utils/FetchWithAuth.tsx";
