+++ src/pages/shoppingPages/ShopMainPage.tsx
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import Header from "../../components/shoppingPages/ShopMainPage/Header.tsx";
 import Main from "../../components/shoppingPages/ShopMainPage/Main.tsx";
 import Nav from "../../components/common/Nav.tsx";
-import useDeviceSize from "../../useDeviceSize.tsx";
+import useDeviceSize from "../../hooks/useDeviceSize.tsx";
 
 //isSmall에 $ 붙이는 이유: styled-components에서 $가 붙은 props는 DOM에 넘기지 않음
 //styled 내부에서만 쓸 값이면 꼭 $ 붙이기
