+++ src/pages/shoppingPages/ShopResultPage.tsx
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import Header from "../../components/shoppingPages/ShopResultPage/Header.tsx";
 import Main from "../../components/shoppingPages/ShopResultPage/Main.tsx";
 import Nav from "../../components/common/Nav.tsx";
-import useDeviceSize from "../../useDeviceSize.tsx";
+import useDeviceSize from "../../hooks/useDeviceSize.tsx";
 import { useLocation } from "react-router-dom";
 import { useEffect, useState } from "react";
 import { fetchWithAuth } from "../../utils/FetchWithAuth.tsx";
